% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_daily_to_weekly.R
\name{aggregate_daily_to_weekly}
\alias{aggregate_daily_to_weekly}
\title{Aggregate daily data to weekly data}
\usage{
aggregate_daily_to_weekly(target_ts)
}
\arguments{
\item{target_ts}{a \code{data.frame} of target data in a time series format
(contains columns \code{time_index}, \code{location}, and \code{observation}) for a single
location}
}
\value{
data.frame of time series data with the same set of input columns, with
weekly-aggregated data in \code{observation} column
}
\description{
Counts weeks as beginning on Sunday and ending on Saturday. Drops observations
from the most recent week if not a full 7 days of data
}
