% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_trends_ensemble.R
\name{create_trends_ensemble}
\alias{create_trends_ensemble}
\title{Generate predictions for the trends ensemble, a quantile median of component
baseline models}
\usage{
create_trends_ensemble(
  component_variations,
  target_ts,
  reference_date,
  horizons,
  target,
  quantile_levels,
  n_samples = NULL,
  round_predictions = FALSE,
  seed = NULL,
  return_baseline_predictions = FALSE
)
}
\arguments{
\item{component_variations}{a \code{data.frame} where each row specifies a set of
hyperparameters to use for a single baseline model fit, with columns
\code{transformation}, \code{symmetrize}, \code{window_size}, and \code{temporal_resolution}.
See details for more information}

\item{target_ts}{a \code{data.frame} of target data in a time series format
(contains columns \code{time_index}, \code{location}, and \code{observation}) for a single
location}

\item{reference_date}{string of the reference date for the forecasts, i.e.
the date relative to which the targets are defined (usually Saturday for
weekly targets). Must be in the ymd format, with yyyy-mm-dd format recommended.}

\item{horizons}{numeric vector of prediction horizons relative to
the reference_date, e.g. 0:3 or 1:4, and interpreted to be in terms of the
same temporal resolution as the provided \code{target_ts}.}

\item{target}{character string specifying the name of the prediction target}

\item{quantile_levels}{numeric vector of quantile levels to output; set to NULL
if quantile outputs not requested. Defaults to NULL.}

\item{n_samples}{integer of amount of samples to output (and predict);
set to NULL if sample outputs not requested (in this case 100000 samples
are generated from which to extract quantiles). Defaults to NULL.}

\item{round_predictions}{boolean specifying whether to round the output
predictions to the nearest whole number. Defaults to FALSE}

\item{seed}{integer specifying a seed to set for reproducible results.
Defaults to NULL, in which case no seed is set.}

\item{return_baseline_predictions}{boolean specifying whether to the component
baseline models' forecasts in addition to the trends ensemble forecasts.
If TRUE, a two-item list will be returned containing a labeled model_out_tbl
of each. Defaults to FALSE.}
}
\value{
\code{model_out_tbl} of trends ensemble forecasts with columns:
\code{model_id}, \code{reference_date}, \code{location}, \code{horizon}, \code{target},
\code{target_end_date}, \code{output_type}, \code{output_type_id}, and \code{value}.
}
\description{
Generate predictions for the trends ensemble, a quantile median of component
baseline models
}
\details{
The \code{component_variations} data frame has the following columns and
possible values for each:
\itemize{
\item transformation (character): "none" or "sqrt", determines distribution shape
\item symmetrize (boolean), determines if distribution is symmetric
\item window_size (integer), determines how many previous observations inform
the forecast
\item temporal_resolution (character): "daily" or "weekly"
}

Note that it must be possible to aggregate the \code{target_ts} data to the
temporal resolution values given in \code{component_variations}. For example, if
\code{target_ts} contains weekly observations but \code{component_variations} requests
models with a "daily" temporal resolution, an error will be thrown
}
