% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_baseline_predictions.R
\name{extract_predictions}
\alias{extract_predictions}
\title{Extract and compute sample and/or quantile forecasts from a prediction matrix}
\usage{
extract_predictions(
  predictions,
  effective_horizons,
  quantile_levels = NULL,
  n_samples = NULL
)
}
\arguments{
\item{predictions}{a \code{matrix} of \code{n_sim} runs of sample predictions with
dimensions n_sim x max(effective_horizons)}

\item{effective_horizons}{numeric vector of prediction horizons relative to
the last observed date in \code{target_ts}}

\item{quantile_levels}{numeric vector of quantile levels to output; set to NULL
if quantile outputs not requested. Defaults to NULL.}

\item{n_samples}{integer number of samples to output, which are drawn from the
simulated (sample) predictions (hence n_samples <= n_sim). Set to NULL
if sample outputs not requested. Defaults to NULL.}
}
\value{
data frame of extracted sample and/or quantile forecasts with
columns \code{horizon} , \code{output_type}, \code{output_type_id}, and \code{value}
}
\description{
Extract and compute sample and/or quantile forecasts from a prediction matrix
}
