% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_baselines_one_location.R
\name{fit_baselines_one_location}
\alias{fit_baselines_one_location}
\title{Generate predictions for all baseline models, for a single location}
\usage{
fit_baselines_one_location(
  model_variations,
  target_ts,
  reference_date,
  temporal_resolution,
  horizons,
  target,
  origin,
  quantile_levels,
  n_samples,
  round_predictions = FALSE,
  seed = NULL
)
}
\arguments{
\item{model_variations}{a \code{data.frame} where each row specifies a set of
hyperparameters to use for a single baseline model fit, with columns
\code{transformation}, \code{symmetrize}, and \code{window_size}. See details for more
information}

\item{target_ts}{a \code{data.frame} of target data in a time series format
(contains columns \code{time_index}, \code{location}, and \code{observation}) for a single
location}

\item{reference_date}{string of the reference date for the forecasts, i.e.
the date relative to which the targets are defined (usually Saturday for
weekly targets)}

\item{temporal_resolution}{'daily' or 'weekly'; specifies timescale of
target_ts and horizons}

\item{horizons}{numeric vector of prediction horizons relative to
the reference_date, e.g. 0:3 or 1:4}

\item{target}{character string specifying the name of the prediction target}

\item{origin}{string specifying the origin to use when making predictions;
recommended to be "median" if the temporal resolution is daily and "obs"
if weekly or otherwise. Defaults to "obs".}

\item{quantile_levels}{numeric vector of quantile levels to output; set to NULL
if quantile outputs not requested. Defaults to NULL.}

\item{n_samples}{integer of amount of samples to output (and predict);
set to NULL if sample outputs not requested (in this case 100000 samples
are generated from which to extract quantiles). Defaults to NULL.}

\item{round_predictions}{boolean specifying whether to round the output
predictions to the nearest whole number. Defaults to FALSE}

\item{seed}{integer specifying a seed to set for reproducible results.
Defaults to NULL, in which case no seed is set.}
}
\value{
data frame of a baseline forecast for one location, all models with
columns \code{transformation}, \code{symmetrize}, \code{window_size}, \code{horizon},
\code{output_type}, \code{output_type_id}, and \code{value}
}
\description{
Generate predictions for all baseline models, for a single location
}
\details{
The \code{model_variations} data frame has the following columns and
possible values for each:
\itemize{
\item transformation (character): "none" or "sqrt", determines distribution shape
\item symmetrize (boolean), determines if distribution is symmetric
\item window_size (integer), determines how many previous observations inform
the forecast
}
}
