% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_baseline_predictions.R
\name{get_baseline_predictions_new}
\alias{get_baseline_predictions_new}
\title{Get predictions for a single baseline model}
\usage{
get_baseline_predictions_new(
  target_ts,
  transformation,
  symmetrize,
  window_size,
  effective_horizons,
  origin = "obs",
  quantile_levels,
  n_samples
)
}
\arguments{
\item{target_ts}{a \code{data.frame} of target data in a time series format
(contains columns \code{time_index}, \code{location}, and \code{observation})}

\item{transformation}{string specifying the transformation used to create
the model; can be one of "none" or "sqrt".}

\item{symmetrize}{boolean specifying whether to symmetrize the model; can be
one of \code{TRUE} or \code{FALSE}.}

\item{window_size}{integer specifying the window size used to create the model}

\item{origin}{string specifying the origin to use when making predictions;
recommended to be "median" if the temporal resolution is daily and "obs"
if weekly or otherwise. Defaults to "obs".}

\item{quantile_levels}{numeric vector of quantile levels to output; set to NA
if quantile outputs not requested}

\item{n_samples}{integer of amount of samples to output (and predict);
set to NA if sample outputs not requested (in this case 100000 samples
are generated from which to extract quantiles)}

\item{effective}{horizons numeric vector of prediction horizons relative to
the last observed date in \code{target_ts}}
}
\value{
data frame of a baseline forecast for one location, one model with
columns \code{horizon} , \code{output_type_id}, and \code{value}, but which are stored
as a nested list in a 1x1 data frame
}
\description{
Get predictions for a single baseline model
}
