% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_baseline_predictions.R
\name{get_baseline_predictions}
\alias{get_baseline_predictions}
\title{Get predictions for a single baseline model}
\usage{
get_baseline_predictions(
  target_ts,
  transformation,
  symmetrize,
  window_size,
  effective_horizons,
  origin = "obs",
  n_sim = 1e+05,
  quantile_levels = NULL,
  n_samples = NULL,
  round_predictions = FALSE,
  seed = NULL
)
}
\arguments{
\item{target_ts}{a \code{data.frame} of target data in a time series format
(contains columns \code{time_index}, \code{location}, and \code{observation}) for a single
location}

\item{transformation}{string specifying the transformation used on the
distribution which determines its shape; can be one of "none" or "sqrt".}

\item{symmetrize}{boolean specifying whether to make the distribution symmetric;
can be one of \code{TRUE} or \code{FALSE}.}

\item{window_size}{integer specifying how many previous observations in the
target data should be used to inform the forecasts}

\item{effective_horizons}{numeric vector of prediction horizons relative to
the last observed date in \code{target_ts}}

\item{origin}{string specifying the origin to use when making predictions;
recommended to be "median" if the temporal resolution is daily and "obs"
if weekly or otherwise. Defaults to "obs".}

\item{n_sim}{integer number of simulations to predict. Defaults to 100000.}

\item{quantile_levels}{numeric vector of quantile levels to output; set to NULL
if quantile outputs not requested. Defaults to NULL.}

\item{n_samples}{integer number of samples to output, which are drawn from the
simulated (sample) predictions (hence n_samples <= n_sim). Set to NULL
if sample outputs not requested. Defaults to NULL.}

\item{round_predictions}{boolean specifying whether to round the output
predictions to the nearest whole number. Defaults to FALSE}

\item{seed}{integer specifying a seed to set for reproducible results.
Defaults to NULL, in which case no seed is set.}
}
\value{
data frame of a baseline forecast for one location, one model with
columns \code{horizon} , \code{output_type}, \code{output_type_id}, and \code{value},
but which are stored as a nested list in a 1x1 data frame
}
\description{
Get predictions for a single baseline model
}
